split logic to container component, prepare for upload logic
